refactor with PublicationFactory::buildfromDoc, refactor findOne method, fix ownership

+++ server/src/publication/publication.service.ts
@@ -1,9 +1,8 @@
 import {
   BadRequestException,
-  ConflictException,
   Injectable,
   InternalServerErrorException,
-  NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { CreatePublicationDto } from './dto/create-publication.dto';
 import { UpdatePublicationDto } from './dto/update-publication.dto';
@@ -11,7 +10,7 @@ import { Publication } from './entities/publication.entity';
 import { publicationFactory } from 'src/factory/publication.factory';
 import { v4 as uuidv4 } from 'uuid';
 import { db, storage } from 'src/database/app.database';
-import { InputFile, Models, Query } from 'node-appwrite';
+import { InputFile, Models } from 'node-appwrite';
 
 @Injectable()
 export class PublicationService {
@@ -20,6 +19,7 @@ export class PublicationService {
     picture: Express.Multer.File,
     authorUid: string,
   ) {
+    let doc: Models.Document;
     const pictureUid = uuidv4();
 
     try {
@@ -39,7 +39,7 @@ export class PublicationService {
       .build();
 
     try {
-      await db.createDocument('DEV', 'PUBLICATIONS', publication.uid, {
+      doc = await db.createDocument('DEV', 'PUBLICATIONS', publication.uid, {
         ...publication.toObject(),
         author: authorUid,
       });
@@ -51,7 +51,7 @@ export class PublicationService {
       }
       throw new BadRequestException(e.message);
     }
-    return publication;
+    return publicationFactory.buildfromDoc(doc);
   }
 
   async findAll() {
@@ -66,51 +66,24 @@ export class PublicationService {
     const publications: Publication[] = [];
 
     docs.documents.forEach((doc: Models.Document) =>
-      publications.push(
-        publicationFactory
-          .setUid(doc['uid'])
-          .setTitle(doc['title'])
-          .setDescription(doc['description'])
-          .setPictureUid(doc['pictureUid'])
-          .setLikes(doc['likes'])
-          .build(),
-      ),
+      publications.push(publicationFactory.buildfromDoc(doc)),
     );
     return publications;
   }
 
   async findOne(publicationId: string) {
-    let docs: Models.DocumentList<Models.Document>;
+    let doc: Models.Document;
 
     try {
-      docs = await db.listDocuments('DEV', 'PUBLICATIONS', [
-        Query.equal('uid', publicationId),
-      ]);
+      doc = await db.getDocument('DEV', 'PUBLICATIONS', publicationId);
     } catch (e) {
       throw new BadRequestException(e.message);
     }
 
-    if (docs.total == 0) {
-      throw new NotFoundException(
-        "Could not find publication with uid '" + publicationId + "'",
-      );
-    }
-    if (docs.total > 1)
-      throw new ConflictException(
-        "Multiple publications with uid '" + publicationId + "'",
-      );
-
-    return publicationFactory
-      .setUid(docs.documents[0]['uid'])
-      .setTitle(docs.documents[0]['title'])
-      .setDescription(docs.documents[0]['description'])
-      .setPictureUid(docs.documents[0]['pictureUid'])
-      .setLikes(docs.documents[0]['likes'])
-      .build();
+    return publicationFactory.buildfromDoc(doc);
   }
 
   async getPicture(pictureId: string) {
-    console.log('pictureIdd: ', pictureId);
     try {
       return storage.getFileView('DEV', pictureId);
     } catch (e) {
@@ -121,10 +94,14 @@ export class PublicationService {
   async update(
     publicationId: string,
     picture: Express.Multer.File,
+    connectedUserUid: string,
     updatePublicationDto: UpdatePublicationDto,
   ) {
     let doc: Models.Document;
     const publication = await this.findOne(publicationId);
+
+    if (publication.author.toString() != connectedUserUid)
+      throw new UnauthorizedException('Only owner can update its publications');
 
     if (picture != undefined || picture != null) {
       const pictureUid = uuidv4();
@@ -144,22 +121,24 @@ export class PublicationService {
     publication.description = updatePublicationDto.description;
 
     try {
-      doc = await db.updateDocument('DEV', 'USERS', publicationId, publication);
+      doc = await db.updateDocument(
+        'DEV',
+        'PUBLICATIONS',
+        publicationId,
+        publication,
+      );
     } catch (e) {
       throw new BadRequestException(e.message);
     }
 
-    return publicationFactory
-      .setUid(doc['uid'])
-      .setTitle(doc['title'])
-      .setDescription(doc['description'])
-      .setPictureUid(doc['pictureUid'])
-      .setLikes(doc['likes'])
-      .build();
+    return publicationFactory.buildfromDoc(doc);
   }
 
-  async remove(publicationId: string) {
+  async remove(publicationId: string, connectedUserUid: string) {
     const publication = await this.findOne(publicationId);
+
+    if (publication.author.toString() != connectedUserUid)
+      throw new UnauthorizedException('Only owner can update its publications');
 
     try {
       await db.deleteDocument('DEV', 'USERS', publicationId);
